move password related functions to logic pkg
